% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{bargraph_BCR_binnedby_benefits}
\alias{bargraph_BCR_binnedby_benefits}
\title{Bar graph of the BCR binned based on benefits}
\usage{
bargraph_BCR_binnedby_benefits(
  results_BCR_national,
  results_BCR_global,
  inputs,
  results_overall
)
}
\arguments{
\item{results_BCR_national}{A dataframe produced from the cost benefit analysis that outlines the national benefit each program}

\item{results_BCR_global}{A dataframe produced from the cost benefit analysis that outlines the global benefit each program}

\item{inputs}{A CSV containing the conservation programs and the P5, P50, and P95 of associated benefits and costs}

\item{results_overall}{A dataframe produced from the cost benefit analysis that outlines the BCR and ranks for each conservation program}
}
\value{
Returns the figure produced
}
\description{
This function can be used when the cost benefit analysis function cba_GplusD_BinnedByBenefit is used. This function creates a figure depicting the
national and global benefits, giving priority to those that have the highest benefit.
}
\examples{
bargraph_BCR_binnedby_benefits(results_BCR_national, results_BCR_global, inputs, results_overall)
}
