% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{bargraph_bin_currentgs}
\alias{bargraph_bin_currentgs}
\title{INSERT}
\usage{
bargraph_bin_currentgs(
  results_benefit_global,
  results_benefit_national,
  inputs,
  results_overall
)
}
\arguments{
\item{results_benefit_global}{TO DO: ADD DESCRIPTION}

\item{results_benefit_national}{TO DO: ADD DESCRIPTION}

\item{inputs}{TO DO: ADD DESCRIPTION}

\item{results_overall}{TO DO: ADD DESCRIPTION}
}
\description{
INSERT
}
\examples{
# STILL TO ADD

}
