% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_benefit_analysis.R
\name{cba_cgain_longtermasp_binnedbyGS}
\alias{cba_cgain_longtermasp_binnedbyGS}
\title{This function performs the cost benefit analysis for an organizations new species prioritization process.}
\usage{
cba_cgain_longtermasp_binnedbyGS(
  org_programs,
  inputs,
  functional_score_max,
  sensitivity
)
}
\arguments{
\item{org_programs}{A vector of the species names included in the analysis}

\item{inputs}{A dataframe containing the organization's different programs costs and benefits for each species}

\item{functional_score_max}{TO DO}

\item{sensitivity}{TO DO}
}
\value{
TO DO
}
\description{
This function performs the cost benefit analysis for an organizations new species prioritization process.
}
