% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_benefit_analysis.R
\name{cba_GplusD_CurrentGS}
\alias{cba_GplusD_CurrentGS}
\title{This function performs the cost benefit analysis for an organizations new species prioritization process.}
\usage{
cba_GplusD_CurrentGS(org_programs, inputs, functional_score_max)
}
\arguments{
\item{org_programs}{a vector of the conservation program names}

\item{inputs}{a CSV containing the conservation programs and the P5, P50, and P95 of associated benefits and costs}

\item{functional_score_max}{the maximum possible score for a spatial unit}
}
\value{
Returns last output of function
}
\description{
This function performs the cost benefit analysis for an organizations new species prioritization process.
}
\examples{
# STILL TO ADD

}
