% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_benefit_analysis.R
\name{cba_GplusD_LongTermPot}
\alias{cba_GplusD_LongTermPot}
\title{Cost benefit analysis with long term potential}
\usage{
cba_GplusD_LongTermPot(org_programs, inputs, functional_score_max, epsilon)
}
\arguments{
\item{org_programs}{A vector of the conservation program names}

\item{inputs}{A CSV containing the conservation programs and the P5, P50, and P95 of associated benefits and costs}

\item{functional_score_max}{The maximum possible score for a spatial unit}

\item{epsilon}{As described in the paper}
}
\value{
Returns a list of dataframes containing all the different results
}
\description{
This function performs the cost benefit analysis where benefits, calculated as Conservation Gains plus Conservation Dependence, are relative to
long-term potential Green Score metric.
}
\examples{
# cba_GplusD_LongTermPot(org_programs,inputs,functional_score_max)
}
