% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_benefit_analysis.R
\name{cba_GplusD_LongTermPot_BinnedByGS}
\alias{cba_GplusD_LongTermPot_BinnedByGS}
\title{Cost benefit analysis with benefits relative to long-term potential, binned by current Green Score status}
\usage{
cba_GplusD_LongTermPot_BinnedByGS(
  org_programs,
  inputs,
  functional_score_max,
  epsilon
)
}
\arguments{
\item{org_programs}{A vector of the conservation program names}

\item{inputs}{A CSV containing the conservation programs and the P5, P50, and P95 of associated benefits and costs}

\item{functional_score_max}{The maximum possible score for a spatial unit}

\item{epsilon}{As described in the paper}
}
\value{
Returns a list of dataframes containing all the different results
}
\description{
This function performs the cost benefit analysis with the benefits, calculated as the Conservation Gain plus Conservation Dependence, relative to
long-term potential of each species. Prior to ranking programs, this function bins conservation programs into groups depending on the species
current Green Score. Species with a lower current Green Score are giving priority and will therefore rank higher, even if the cost benefit ratio
is comparatively lower.
}
\examples{
# cba_GplusD_LongTermPot_BinnedByGS(org_programs, inputs, functional_score_max)
}
