% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_benefit_analysis.R
\name{cba_GplusD_LongTermPot_CurrentGS_epsilon}
\alias{cba_GplusD_LongTermPot_CurrentGS_epsilon}
\title{Cost benefit analysis with benefits relative to both long term potential and current Green Score plus epsilon}
\usage{
cba_GplusD_LongTermPot_CurrentGS_epsilon(
  org_programs,
  inputs,
  functional_score_max,
  epsilon
)
}
\arguments{
\item{org_programs}{A vector of the conservation program names}

\item{inputs}{A CSV containing the conservation programs and the P5, P50, and P95 of associated benefits and costs}

\item{functional_score_max}{The maximum possible score for a spatial unit}

\item{epsilon}{As described in the paper}
}
\value{
Returns a list of dataframes containing all the different results
}
\description{
This function performs the cost benefit analysis using an advanced method wherein the benefits, calculated as Conservation Gain plus Conservation
Dependence, are relative to both the long term potential and current the Green Score of the species while also incorporating epsilon. Epsilon is
subjective value that is added to the current Green Score in order to accommodate species that have a current Green Score of zero.
}
\examples{
# cba_GplusD_LongTermPot_CurrentGS_epsilon(org_programs,inputs,functional_score_max)
}
