% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{draw_tornados}
\alias{draw_tornados}
\title{STILL TO CLEAN UP CODE
INSERT}
\usage{
draw_tornados(inputs_sens)
}
\arguments{
\item{inputs_sens}{TO DO: ADD DESCRIPTION}
}
\description{
STILL TO CLEAN UP CODE
INSERT
}
\examples{
# STILL TO ADD

}
