% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{draw_tornados_national}
\alias{draw_tornados_national}
\title{Draws the tornado plots from the sensitivity analysis}
\usage{
draw_tornados_national(inputs_sens)
}
\arguments{
\item{inputs_sens}{A dataframe generated from the sensitivity function}
}
\description{
This function takes the output from the sensitivity analysis and creates tornado plots of the recalculated results at the national level,
highlighting the effects that altering each parameter would have on the ranking of the programs.
}
\examples{
# draw_tornados_national(inputs_sens)
}
