% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataset.R
\docType{data}
\name{ExampleDataset}
\alias{ExampleDataset}
\title{Example dataset for newSpeciesPrioritization package}
\format{
A dataframe with 72 rows and 14 columns:
\describe{
\item{species}{Project or Species name}
\item{category}{Benefits or Costs}
\item{subcategory}{Type of benefit or cost}
\item{type}{Description of subcategory}
\item{unit}{Unit of measure}
\item{name}{Unique identifer of species and measurement}
\item{lowP5}{Low estimate}
\item{baseP50}{Best guess estimate}
\item{highP95}{High estimate}
\item{lowGeneral}{Low estimate of percent of species range within country of interest}
\item{lowGeneralprob}{Probability of associated range}
\item{highGeneral}{High estimate of percent of species range within country of interest}
\item{highGeneralprob}{Probability of associated range}
}
}
\source{
Example data, no academic source
}
\usage{
ExampleDataset
}
\description{
This dataset shows the necessary layout and components of the inputs.csv file that is imported
}
\keyword{datasets}
