% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{graph_BCR}
\alias{graph_BCR}
\title{STILL TO CLEAN UP CODE
INSERT}
\usage{
graph_BCR(results_overall, inputs)
}
\arguments{
\item{results_overall}{TO DO: ADD DESCRIPTION}

\item{inputs}{TO DO: ADD DESCRIPTION}
}
\description{
STILL TO CLEAN UP CODE
INSERT
}
\examples{
# STILL TO ADD

}
