% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{graph_BCR_directional_uncertainty}
\alias{graph_BCR_directional_uncertainty}
\title{Figure of the directional uncertainty of the mean BCR results}
\usage{
graph_BCR_directional_uncertainty(
  results_BCR_national,
  results_BCR_global,
  inputs
)
}
\arguments{
\item{results_BCR_national}{A dataframe produced from the cost benefit analysis that outlines BCR and ranks for each conservation program at the national level}

\item{results_BCR_global}{A dataframe produced from the cost benefit analysis that outlines BCR and ranks for each conservation program at the global level}

\item{inputs}{A CSV containing the conservation programs and the P5, P50, and P95 of associated benefits and costs}
}
\value{
Returns the figure produced
}
\description{
This function produces a figure of the directional uncertainty around the zero-centered means benefit-cost ratios at the national and global ratios for each program.
}
\examples{
graph_BCR_uncertainty(results_BCR_national, results_BCR_global, inputs)
}
