% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{graph_benefit}
\alias{graph_benefit}
\title{STILL TO CLEAN UP CODE
INSERT}
\usage{
graph_benefit(results_benefit_national, results_benefit_global, inputs)
}
\arguments{
\item{results_benefit_national}{TO DO: ADD DESCRIPTION}

\item{results_benefit_global}{TO DO: ADD DESCRIPTION}

\item{inputs}{TO DO: ADD DESCRIPTION}
}
\description{
STILL TO CLEAN UP CODE
INSERT
}
\examples{
# STILL TO ADD

}
