% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_prep_prework.R
\docType{data}
\name{inputs}
\alias{inputs}
\title{Create a data frame with the number of iterations as rows and the inputs as columns.
Populate with randomly drawn values from the relevant distributions for each uncertain input.
Also does some preliminary calculations on costs and includes the results as they are used
for calculating the fundability parameter draws.}
\format{
An object of class \code{data.frame} with 192 rows and 15 columns.
}
\usage{
inputs
}
\description{
Create a data frame with the number of iterations as rows and the inputs as columns.
Populate with randomly drawn values from the relevant distributions for each uncertain input.
Also does some preliminary calculations on costs and includes the results as they are used
for calculating the fundability parameter draws.
}
\keyword{datasets}
