% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsp.R
\name{nsp}
\alias{nsp}
\title{Generic NSP function.}
\usage{
nsp(x, constraints, M, thresh, overlap = FALSE, buffer = 0)
}
\arguments{
\item{x}{data (referred to in the paper as Y).}

\item{constraints}{design matrix (referred to in the paper as X).}

\item{M}{number of intervals to draw; this implementation uses a deterministic equispaced grid for drawing intervals.}

\item{thresh}{threshold value (lambda_alpha in the paper).}

\item{overlap}{FALSE means no overlap, TRUE means an overlap. See Section 4 of the paper.}

\item{buffer}{set to zero if no autoregression present in constraints; otherwise set to the order of the autoregression.}
}
\value{
object of a class with the following two fields -
intervals - data frame whose first two columns are start- and end-points (respectively) of the detected intervals of significance;
   the third column are the corresponding dyadic multiresolution sup-norms.
threshold.used - thresh.
}
\description{
Do not use unless you know what 'thresh' value to set. Use one of the nsp_* functions below (without _selfnorm) instead.
}
