% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_prep_prework.R
\name{parameter_prep}
\alias{parameter_prep}
\title{Prepares data from the inputs csv file and draws values for each parameter}
\usage{
parameter_prep(inputs, org_program, G_and_A_prop, number_of_simulations)
}
\arguments{
\item{inputs}{A CSV containing the conservation programs and the P5, P50, and P95 of associated benefits and costs}

\item{org_program}{A vector of the conservation program names}

\item{G_and_A_prop}{A value specifying the proportion of total cost that must be added on top of total cost}

\item{number_of_simulations}{A value specifying number of value draws for each parameter for each conservation program}
}
\value{
A dataframe "dat" for a single conservation program which contains n number of values (determined by number_of_simulations argument) for the costs and benefit metrics
}
\description{
This function prepares the inputs.csv file, creates the metalog distributions for each parameter using the P5, P50, and P95, and
lastly draws values from the distribution for parameter. The number of values that are generated is dependent on the user-specified
"number_of_simulations"
}
\examples{
parameter_prep(inputs,org_program, G_and_A_prop, number_of_simulations)
}
