% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{scatter_bin_benefits_national}
\alias{scatter_bin_benefits_national}
\title{Scatter plot of national benefits binned by benefits}
\usage{
scatter_bin_benefits_national(
  results_benefit_national,
  inputs,
  results_overall
)
}
\arguments{
\item{results_benefit_national}{A dataframe produced from the cost benefit analysis that outlines the national benefit each program}

\item{inputs}{A CSV containing the conservation programs and the P5, P50, and P95 of associated benefits and costs}

\item{results_overall}{A dataframe produced from the cost benefit analysis that outlines the BCR and ranks for each conservation program}
}
\value{
Returns the figure produced
}
\description{
This function can be used when the cost benefit analysis function cba_GplusD_BinnedByBenefit is used. This function creates a scatter plot of the
national benefits of each  program, binned into groups by the benefits they provide (high, medium, low, zero)
}
\examples{
#scatter_bin_benefits_national(results_benefit_national, inputs, results_overall)
}
