% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{scatter_bin_cgains_global}
\alias{scatter_bin_cgains_global}
\title{INSERT}
\usage{
scatter_bin_cgains_global(results_benefit_global, inputs, results_overall)
}
\arguments{
\item{results_benefit_global}{TO DO: ADD DESCRIPTION}

\item{inputs}{TO DO: ADD DESCRIPTION}

\item{results_overall}{TO DO: ADD DESCRIPTION}

\item{results_benefit_national}{TO DO: ADD DESCRIPTION}
}
\description{
INSERT
}
\examples{
# STILL TO ADD

}
