% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{scatter_bin_currentgs_national}
\alias{scatter_bin_currentgs_national}
\title{INSERT}
\usage{
scatter_bin_currentgs_national(results_overall, results_benefit_national)
}
\arguments{
\item{results_overall}{TO DO: ADD DESCRIPTION}

\item{results_benefit_national}{TO DO: ADD DESCRIPTION}
}
\description{
INSERT
}
\examples{
# STILL TO ADD

}
