% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{sensitivity}
\alias{sensitivity}
\title{Performs the sensitivity analysis on the results of the cost benefit analysis}
\usage{
sensitivity(org_programs, inputs, results_full_analysis, rank_cutoff)
}
\arguments{
\item{org_programs}{A vector of the conservation program names}

\item{inputs}{A CSV containing the conservation programs and the P5, P50, and P95 of associated benefits and costs}

\item{results_full_analysis}{A list of results generated by the cost benefit analysis function}

\item{rank_cutoff}{A value set by user to highlight their organization's number of conservation programs of interest}
}
\value{
Returns a dataframe of the recalculated values
}
\description{
This function takes the output from the cost benefit analysis and the inputs csv file and holds each parameter for each species program
at a high or low value and then recalculates values and performs the ranking again. This aims to provide insight into which parameters may be of
interest.
}
\examples{
sensitivity(org_programs, inputs, results_full_analysis, rank_cutoff)
}
