% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{graph_BCR}
\alias{graph_BCR}
\title{Figure of the benefit-cost ratios}
\usage{
graph_BCR(results_BCR_national, results_BCR_global, inputs, endemic_species)
}
\arguments{
\item{results_BCR_national}{a dataframe produced from the cost benefit analysis that outlines BCR and ranks for each conservation program at the national level}

\item{results_BCR_global}{a dataframe produced from the cost benefit analysis that outlines BCR and ranks for each conservation program at the global level}

\item{inputs}{A CSV containing the conservation programs and the P5, P50, and P95 of associated benefits and costs}

\item{endemic_species}{A vector specifying which species are endemic (e.g., c("Species 2","Species 4")}
}
\value{
Returns the figure produced
}
\description{
This function produces a bargraph of the national and global benefit-cost ratios calculated in the analysis functions.
}
\examples{
#graph_BCR(results_BCR_national, results_BCR_global)
}
