% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{graph_benefit}
\alias{graph_benefit}
\title{Figure of national and global benefits}
\usage{
graph_benefit(
  results_benefit_national,
  results_benefit_global,
  inputs,
  endemic_species
)
}
\arguments{
\item{results_benefit_national}{A dataframe produced from the cost benefit analysis that outlines the national benefit each program}

\item{results_benefit_global}{A dataframe produced from the cost benefit analysis that outlines the global benefit each program}

\item{inputs}{A CSV containing the conservation programs and the P5, P50, and P95 of associated benefits and costs}

\item{endemic_species}{A vector specifying which species are endemic (e.g., c("Species 2","Species 4")}
}
\value{
Returns the figure produced
}
\description{
This function creates a bar graph of the national and global benefits with programs that have a current Green Score of zero on a separate facet.
}
\examples{
#graph_benefit(results_benefit_national, results_benefit_global, inputs)
}
