% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{graph_cost}
\alias{graph_cost}
\title{STILL TO CLEAN UP CODE
INSERT}
\usage{
graph_cost(results_cost_organization, results_cost_total, inputs)
}
\arguments{
\item{results_cost_organization}{TO DO: ADD DESCRIPTION}

\item{results_cost_total}{TO DO: ADD DESCRIPTION}

\item{inputs}{TO DO: ADD DESCRIPTION}
}
\description{
STILL TO CLEAN UP CODE
INSERT
}
\examples{
# STILL TO ADD

}
