% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{graph_cost}
\alias{graph_cost}
\title{Figure of costs to organization and financial leverage}
\usage{
graph_cost(results_cost_organization, results_cost_total, inputs)
}
\arguments{
\item{results_cost_organization}{A dataframe produced from the cost benefit analysis that outlines the cost of each program to the organization}

\item{results_cost_total}{A dataframe produced from the cost benefit analysis that outlines the total cost of each program (including to organization and the amount offset by funders)}

\item{inputs}{A CSV containing the conservation programs and the P5, P50, and P95 of associated benefits and costs}
}
\value{
Returns the figure produced
}
\description{
This function creates a figure depicting the costs to the organization and the financial leverage for each conservation program. Programs
with a current Green Score of zero are presented on a separate facet.
}
\examples{
#graph_cost(results_cost_organization, results_cost_total, inputs)
}
