% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{scatter_bin_currentgs_global}
\alias{scatter_bin_currentgs_global}
\title{Scatter plot of the global benefits binned by current Green Score}
\usage{
scatter_bin_currentgs_global(results_overall, results_benefit_global)
}
\arguments{
\item{results_overall}{A dataframe produced from the cost benefit analysis that outlines BCR and ranks for each conservation program}

\item{results_benefit_global}{A dataframe produced from the cost benefit analysis that outlines the global benefit of each program}
}
\value{
Returns the figure produced
}
\description{
This function can be used when the cost benefit analysis function cba_GplusD_LongTermPot_BinnedByGS is used. This function creates a scatter plot
of the global benefits of each conservation program, binned by the current Green Score.
}
\examples{
#scatter_bin_currentgs_global(results_overall, results_benefit_global)
}
