% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{STILL TO CLEAN UP CODE
INSERT}
\usage{
sensitivity(org_programs, inputs, results_full_analysis, rank_cutoff)
}
\arguments{
\item{org_programs}{TO DO: ADD DESCRIPTION}

\item{inputs}{TO DO: ADD DESCRIPTION}

\item{results_full_analysis}{TO DO: ADD DESCRIPTION}

\item{rank_cutoff}{TO DO: ADD DESCRIPTION, FOR US IT IS 6 BECAUSE WE WANT THE TOP 6 SPECIES}
}
\description{
STILL TO CLEAN UP CODE
INSERT
}
\examples{
# STILL TO ADD

}
